+++ Sources/Module/MessageSearch/SBUMessageSearchModule.List.swift
@@ -133,6 +133,7 @@ extension SBUMessageSearchModule {
             self.tableView.backgroundView = self.emptyView
             self.tableView.rowHeight = UITableView.automaticDimension
             self.tableView.estimatedRowHeight = 44.0
+            self.tableView.sectionHeaderHeight = 0
             self.addSubview(self.tableView)
             
             // register cell
@@ -240,6 +241,14 @@ extension SBUMessageSearchModule.List: SBUEmptyViewDelegate {
 
 // MARK: - UITableView relations
 extension SBUMessageSearchModule.List: UITableViewDataSource, UITableViewDelegate {
+    open func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        nil
+    }
+
+    open func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        0
+    }
+    
     open func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return self.resultList.count
     }