+++ Sources/Module/SelectUser/SBUBaseSelectUserModule.List.swift
@@ -104,6 +104,7 @@ extension SBUBaseSelectUserModule {
             self.tableView.backgroundView = self.emptyView
             self.tableView.rowHeight = UITableView.automaticDimension
             self.tableView.estimatedRowHeight = 44.0
+            self.tableView.sectionHeaderHeight = 0
             self.addSubview(self.tableView)
             
             // register cell
@@ -205,6 +206,14 @@ extension SBUBaseSelectUserModule.List: SBUEmptyViewDelegate {
 
 // MARK: - UITableView relations
 extension SBUBaseSelectUserModule.List: UITableViewDataSource, UITableViewDelegate {
+    open func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        nil
+    }
+
+    open func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        0
+    }
+    
     open func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return self.userList?.count ?? 0
     }