+++ Sources/Module/UserList/SBUUserListModule.List.swift
@@ -153,6 +153,7 @@ extension SBUUserListModule {
             self.tableView.backgroundView = self.emptyView
             self.tableView.rowHeight = UITableView.automaticDimension
             self.tableView.estimatedRowHeight = 44.0
+            self.tableView.sectionHeaderHeight = 0
             self.addSubview(self.tableView)
             
             // register cell
@@ -328,6 +329,14 @@ extension SBUUserListModule.List: SBUEmptyViewDelegate {
 
 // MARK: - UITableView relations
 extension SBUUserListModule.List: UITableViewDataSource, UITableViewDelegate {
+    open func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        nil
+    }
+
+    open func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        0
+    }
+    
     open func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return self.userList.count
     }