+++ Sources/View/SelectUser/SBUBaseSelectUserViewController.swift
@@ -99,7 +99,7 @@ open class SBUBaseSelectUserViewController: SBUBaseViewController, SBUBaseSelect
     // MARK: - Actions
     
     /// This function is used to pop to channelViewController.
-    public func popToChannel() {
+    open func popToChannel() {
         guard let navigationController = self.navigationController,
               navigationController.viewControllers.count > 1 else {
                   self.dismiss(animated: true, completion: nil)